Refactor: rename write_pot to write_elecstat_pot
