+++ source/Makefile.Objects
@@ -491,7 +491,7 @@ OBJS_IO=input_conv.o\
     write_wfc_pw.o\
     winput.o\
     write_cube.o\
-    write_pot.o\
+    write_elecstat_pot.o\
     write_dipole.o\
     td_current_io.o\
     write_wfc_r.o\
