+++ source/module_esolver/esolver_ks_lcao.cpp
@@ -12,8 +12,8 @@
 #include "module_io/to_qo.h"
 #include "module_io/write_HS.h"
 #include "module_io/write_eband_terms.hpp"
+#include "module_io/write_elecstat_pot.h"
 #include "module_io/write_istate_info.h"
-#include "module_io/write_pot.h"
 #include "module_io/write_proj_band_lcao.h"
 #include "module_io/write_vxc.hpp"
 #include "module_parameter/parameter.h"
