+++ source/module_esolver/esolver_ks_pw.cpp
@@ -37,7 +37,7 @@
 #include "module_io/numerical_descriptor.h"
 #include "module_io/to_wannier90_pw.h"
 #include "module_io/winput.h"
-#include "module_io/write_pot.h"
+#include "module_io/write_elecstat_pot.h"
 #include "module_io/write_wfc_r.h"
 #include "module_parameter/parameter.h"
 #ifdef USE_PAW
