+++ source/module_esolver/esolver_of.cpp
@@ -2,7 +2,7 @@
 
 #include "module_io/cube_io.h"
 #include "module_io/output_log.h"
-#include "module_io/write_pot.h"
+#include "module_io/write_elecstat_pot.h"
 //-----------temporary-------------------------
 #include "module_base/global_function.h"
 #include "module_elecstate/module_charge/symmetry_rho.h"
